# api/docs/swagger.yaml
basePath: /auth
definitions:
  models.AcceptCode:
    properties:
      code:
        type: string
      email:
        type: string
    type: object
  models.Error:
    properties:
      error:
        type: string
    type: object
  models.ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
  models.LoginEmailRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.LoginUsernameRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.Message:
    properties:
      message:
        type: string
    type: object
  models.RegisterRequest:
    properties:
      bio:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      nationality:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  models.RegisterResponse:
    properties:
      bio:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      nationality:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  models.ResetPassReq:
    properties:
      code:
        type: string
      email:
        type: string
      new_password:
        type: string
    type: object
  models.Tokens:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
info:
  contact: {}
  description: server for siginIn or signUp
  title: Authenfication service
  version: "1.0"
paths:
  /accept-code:
    post:
      description: it accepts code to register
      parameters:
      - description: enough
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/models.AcceptCode'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Accept code to register
      tags:
      - Auth
  /forgot-password:
    post:
      description: it sends code to your email address
      parameters:
      - description: enough
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/models.ForgotPasswordRequest'
      responses:
        "200":
          description: message
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Forgot Password
      tags:
      - Auth
  /login/email:
    post:
      consumes:
      - application/json
      description: sign in user
      parameters:
      - description: register user
        in: body
        name: LoginEmail
        required: true
        schema:
          $ref: '#/definitions/models.LoginEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tokens'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: LoginEmail Users
      tags:
      - Auth
  /login/username:
    post:
      consumes:
      - application/json
      description: sign in user
      parameters:
      - description: register user
        in: body
        name: LoginUsername
        required: true
        schema:
          $ref: '#/definitions/models.LoginUsernameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tokens'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: LoginUsername Users
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      description: create users
      parameters:
      - description: register user
        in: body
        name: Register
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Register Users
      tags:
      - Auth
  /register-admin:
    post:
      description: Registers a new user`
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Registers user
      tags:
      - Auth
  /reset-password:
    post:
      description: it Reset your Password
      parameters:
      - description: enough
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/models.ResetPassReq'
      responses:
        "200":
          description: message
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Reset Password
      tags:
      - Auth
schemes:
- http
swagger: "2.0"